Refactor route and stop pages into separate activities
