+++ app/build.gradle.kts
@@ -43,22 +43,22 @@ android {
         buildConfig = true
     }
 
-    secrets {
-        // To add your Maps API key to this project:
-        // 1. If the secrets.properties file does not exist, create it in the same folder as the local.properties file.
-        // 2. Add this line, where YOUR_API_KEY is your API key:
-        //        MAPS_API_KEY=YOUR_API_KEY
-        propertiesFileName = "secrets.properties"
-
-        // A properties file containing default secret values. This file can be
-        // checked in version control.
-        defaultPropertiesFileName = "local.defaults.properties"
-
-        // Configure which keys should be ignored by the plugin by providing regular expressions.
-        // "sdk.dir" is ignored by default.
-        ignoreList.add("keyToIgnore") // Ignore the key "keyToIgnore"
-        ignoreList.add("sdk.*")       // Ignore all keys matching the regexp "sdk.*"
-    }
+//    secrets {
+//        // To add your Maps API key to this project:
+//        // 1. If the secrets.properties file does not exist, create it in the same folder as the local.properties file.
+//        // 2. Add this line, where YOUR_API_KEY is your API key:
+//        //        MAPS_API_KEY=YOUR_API_KEY
+//        propertiesFileName = "secrets.properties"
+//
+//        // A properties file containing default secret values. This file can be
+//        // checked in version control.
+//        defaultPropertiesFileName = "local.defaults.properties"
+//
+//        // Configure which keys should be ignored by the plugin by providing regular expressions.
+//        // "sdk.dir" is ignored by default.
+//        ignoreList.add("keyToIgnore") // Ignore the key "keyToIgnore"
+//        ignoreList.add("sdk.*")       // Ignore all keys matching the regexp "sdk.*"
+//    }
 }
 
 dependencies {
@@ -74,7 +74,7 @@ dependencies {
     implementation(libs.androidx.ui.text.google.fonts)
     implementation(libs.play.services.maps)
     implementation(libs.androidx.appcompat)
-    implementation(libs.places)
+    //implementation(libs.places)
     testImplementation(libs.junit)
     androidTestImplementation(libs.androidx.junit)
     androidTestImplementation(libs.androidx.espresso.core)
