+++ app/src/main/java/com/example/watcomtravels/WTAApi.kt
@@ -5,11 +5,11 @@ import android.os.Parcel
 import android.os.Parcelable
 import android.util.Log
 import com.google.android.gms.maps.model.LatLng
-import com.google.android.libraries.places.api.Places
-import com.google.android.libraries.places.api.model.Place
-import com.google.android.libraries.places.api.model.RectangularBounds
-import com.google.android.libraries.places.api.net.PlacesClient
-import com.google.android.libraries.places.api.net.SearchByTextRequest
+//import com.google.android.libraries.places.api.Places
+//import com.google.android.libraries.places.api.model.Place
+//import com.google.android.libraries.places.api.model.RectangularBounds
+//import com.google.android.libraries.places.api.net.PlacesClient
+//import com.google.android.libraries.places.api.net.SearchByTextRequest
 import com.google.maps.android.ktx.BuildConfig
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
@@ -22,6 +22,7 @@ import org.json.JSONObject
 import org.json.JSONStringer
 import java.io.FileNotFoundException
 import java.io.IOException
+import java.io.Serializable
 import java.net.HttpURLConnection
 import java.net.URL
 import java.util.Arrays
@@ -455,6 +456,30 @@ class WTAApi {
 
         }
 
+        // Gets the stop information through the route Number
+        fun getRoute(routeNum: String) : Route? {
+            val json = callAPI("https://api.ridewta.com/routes/$routeNum")
+
+            if (json == null) {
+                return null
+            } else {
+                val jsonObject = JSONObject(json)
+
+                val name = jsonObject.getString("routeName")
+                val color = jsonObject.getString("routeColor")
+                val pattern = getRoutePatterns(routeNum)
+
+                val route = Route(
+                    routeNum = routeNum,
+                    name = name,
+                    color = color,
+                    pattern = pattern
+                )
+
+                return route
+            }
+        }
+
         fun getAllBulletins(): List<ServiceBulletin>{
             val bulletinList = mutableListOf<ServiceBulletin>()
 
@@ -512,38 +537,38 @@ class WTAApi {
         /**
          * Calls the placesAPI
          */
-        private fun callPlacesAPI(str: String) : String? {
-            var retStr : String
-            try {
-                // Specify what kind of things to return
-                val placeFields : List<Place.Field> = Arrays.asList(Place.Field.ID, Place.Field.DISPLAY_NAME)
-
-                // Lat Long bounds for the search
-                val swBound = LatLng(48.40004, -122.37137)
-                val neBound = LatLng(48.51422, -122.19048)
-
-                val searchByTextRequest = SearchByTextRequest.builder(str, placeFields)
-                    .setMaxResultCount(10)
-                    .setLocationRestriction(RectangularBounds.newInstance(swBound, neBound)).build()
-
-
-            } catch (e : Exception) {
-                Log.d("@_@", "Error: $e")
-                return "??"
-            }
-            return ""
-        }
+//        private fun callPlacesAPI(str: String) : String? {
+//            var retStr : String
+//            try {
+//                // Specify what kind of things to return
+//                val placeFields : List<Place.Field> = Arrays.asList(Place.Field.ID, Place.Field.DISPLAY_NAME)
+//
+//                // Lat Long bounds for the search
+//                val swBound = LatLng(48.40004, -122.37137)
+//                val neBound = LatLng(48.51422, -122.19048)
+//
+//                val searchByTextRequest = SearchByTextRequest.builder(str, placeFields)
+//                    .setMaxResultCount(10)
+//                    .setLocationRestriction(RectangularBounds.newInstance(swBound, neBound)).build()
+//
+//
+//            } catch (e : Exception) {
+//                Log.d("@_@", "Error: $e")
+//                return "??"
+//            }
+//            return ""
+//        }
 
 
         /**
          * Test function for places API
          */
-        fun getPlacesSearch(str: String) : String? {
-            var json : String
-            runBlocking {
-                json = callPlacesAPI(str).toString()
-            }
-            return json
-        }
+//        fun getPlacesSearch(str: String) : String? {
+//            var json : String
+//            runBlocking {
+//                json = callPlacesAPI(str).toString()
+//            }
+//            return json
+//        }
     }
 }
